requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.3.0  # Updated for Python 3.13 compatibility
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Segmentation and Detection
# detectron2>=0.6.0  # Install with: pip install 'git+https://github.com/facebookresearch/detectron2.git'
# mmcv>=2.0.0  # Install separately due to build issues
# mmdet>=3.0.0  # Install separately after mmcv
# mmsegmentation>=1.0.0  # Install separately after mmcv

# NLP and ASR
google-generativeai>=0.7.2
openai-whisper>=20231117
sentence-transformers>=2.2.0
transformers>=4.30.0
torchaudio>=2.0.0  # Fixed typo: was torch-audio
librosa>=0.10.0

# Optimization
cvxpy>=1.3.0
ortools>=9.6.0
scipy>=1.11.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
pycocotools>=2.0.6
h5py>=3.9.0
tqdm>=4.65.0

# API and Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
gunicorn>=21.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
boto3>=1.28.0

# Monitoring and Logging
mlflow>=2.5.0
wandb>=0.15.0
structlog>=23.0.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Edge Deployment
onnx>=1.14.0
onnxruntime>=1.15.0
# tensorrt>=8.6.0  # For Jetson deployment - not available on Windows
# tflite-runtime>=2.13.0  # For EdgeTPU - not available on Windows

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Geospatial (for GPS and mapping)
geopy>=2.3.0
folium>=0.14.0
shapely>=2.0.0

# MAVLink and UAV
pymavlink>=2.4.37
dronekit>=2.9.2
